Improved error handling
Improvement in error handling in fetch_tunnels. Now, when an error occurs in the HTTP request or the response is not 200, None is returned.

In async_update_data, it checks if the received data is None and in that case, an UpdateFailed exception is thrown with an appropriate message.

In the CloudflareTunnelSensor class, in the async_update method, it now checks if self.coordinator.data is None before trying to process the data:
1. If there is no data, an error is logged.
2. If there is data, the updated tunnel is looked for in the tunnel list.
3. If the tunnel is not found, an error is logged.
Added more debug logs (_LOGGER.debug) to facilitate tracking of data and program flow, which is useful for identifying problems.